pyproject.toml
[tool.poetry]
name = "project-scraper"
version = "0.1.0"
description = ""
authors = ["melanylibia <elibialany@gmail.com>"]
readme = "README.md"
packages = [{include = "project_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
scrapy = "^2.12.0"
twisted = "^24.10.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
